Make LinearCombination subclss of Sum
