+++ pennylane/ops/op_math/__init__.py
@@ -129,8 +129,7 @@ from .exp import Exp, exp
 from .pow import Pow, pow
 from .prod import Prod, prod
 from .sprod import SProd, s_prod
-from .sum import Sum, sum
-from .linear_combination import LinearCombination
+from .sum import Sum, sum, LinearCombination
 from .symbolicop import ScalarSymbolicOp, SymbolicOp
 from .controlled_decompositions import ctrl_decomp_zyz, ctrl_decomp_bisect
